refactor(tui): Extract HelpView to views/HelpView.tsx (#1582)
Move HelpView component (177 lines) from app.tsx to its own file:

- Create views/HelpView.tsx with HelpView and ShortcutRow components
- Update app.tsx to import HelpView from new location
- Remove unused useInput import from app.tsx
- Fix ESLint array type error

This improves code organization and allows HelpView to be tested
in isolation, consistent with other views in /views directory.
